Bug 1484012: [fetch-content] Add an option to not unpack downloaded artifacts; r=gps
Differential Revision: https://phabricator.services.mozilla.com/D3554

--HG--
extra : rebase_source : 58bba31bd921d29d4a40ad8d9ba09c4c7ac1f8dc

+++ testing/mozharness/mozharness/mozilla/fetches.py
@@ -56,5 +56,6 @@ class FetchesMixin(object):
 
             path = self.download_file(url, parent_dir=extdir)
 
-            mozfile.extract(path, extdir)
-            os.remove(path)
+            if fetch['extract']:
+                mozfile.extract(path, extdir)
+                os.remove(path)
